[BlackBerry] Autofill feature implementation for BlackBerry porting https://bugs.webkit.org/show_bug.cgi?id=85577
Patch by Jonathan Dong <jonathan.dong@torchmobile.com.cn> on 2012-06-11
Reviewed by Rob Buis.

Remove the autofill saving procedure out of dispatchWillSubmitForm, this procedure
is same as the procedure in dispatchWillSendSubmitEvent.
Some sites will redirect to other url in its login process, which will cause the
user get notified to save credentials for this provisional redirecting url if we
put this saving procedure in dispatchWillSubmitForm. So we should remove it to make
sure we only save the autofill data before the submit event is fired.

* WebCoreSupport/FrameLoaderClientBlackBerry.cpp:
(WebCore::FrameLoaderClientBlackBerry::dispatchWillSubmitForm):

Identifier: 106655@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120022 268f45cc-cd09-0410-ab3c-d52691b4dbfc