Event dispatcher should use InsertionPoint::hasDistribution instead of InsertinPoint::isActive in re-targeting. https://bugs.webkit.org/show_bug.cgi?id=89180
Reviewed by Dimitri Glazkov.

Source/WebCore:

If an event is fired on a fallback element within a insertion
point which does not select any nodes, the result of event's
re-targeting will be wrong since it uses
InsertionPoint::isActive() to judge an existence of a lower boundary.
That must use InsertionPoint::hasDistribution().

Test: fast/dom/shadow/shadow-dom-event-dispatching.html

* dom/EventDispatcher.cpp:
(WebCore::EventDispatcher::ensureEventAncestors):

LayoutTests:

* fast/dom/shadow/shadow-dom-event-dispatching-expected.txt:
* fast/dom/shadow/shadow-dom-event-dispatching.html:


Identifier: 107227@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120615 268f45cc-cd09-0410-ab3c-d52691b4dbfc