Fix regression introduced in r120646 where the svn revision isn't saved in results.json
Unreviewed, build fix.

It turns out that we store the svn revision two different ways
in the JSON, and I only caught one of them :(. This should fix
the other.

Also, this part of the code is poorly tested. I will investigate
adding at least a basic test for this in a separate fix.

* Scripts/webkitpy/layout_tests/controllers/manager.py:
(summarize_results):

Identifier: 107349@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120743 268f45cc-cd09-0410-ab3c-d52691b4dbfc