Refine syncLayersTimeoutCallback for Accelerated Compositing. https://bugs.webkit.org/show_bug.cgi?id=89538
Patch by Chang Wan Hong <jourmoon@company100.net> on 2012-06-19
Reviewed by Martin Robinson.

syncLayersTimeout reschedules the timer so that it can render each frame
every 1/60 seconds on animation. However, because it takes to time to execute
renderLayersToWindow, the timer is delayed. To fix this, we must reschedule
the timer before calling renderLayersToWindow.

* WebCoreSupport/AcceleratedCompositingContextGL.cpp:
(WebKit::AcceleratedCompositingContext::syncLayersTimeout):

Identifier: 107405@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120803 268f45cc-cd09-0410-ab3c-d52691b4dbfc