[chromium] Select the marker range when right-clicking on a marker. https://bugs.webkit.org/show_bug.cgi?id=89331
Reviewed by Hajime Morita.

Chromium always selects only one word when right-clicking on a spelling marker.
This prevents selecting whole region specified by a marker if the marker
consists of two or more words. This change retrieves a range coverted by a
marker and select the range. This emulates the behavior of Mac Chromium and it
does not need any workarounds for Mac.

* src/ContextMenuClientImpl.cpp:
(WebKit::ContextMenuClientImpl::getCustomMenuFromDefaultItems):


Identifier: 107412@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120810 268f45cc-cd09-0410-ab3c-d52691b4dbfc