[Shadow][Editing] Selection will break editing boundaries in Shadow DOM. https://bugs.webkit.org/show_bug.cgi?id=89075
Reviewed by Antti Koivisto.

Source/WebCore:

At the shadow boundary, some styles (e.g. user-modify, text-decoration) cannot be inherited
from the shadow host. However, when style property cache is used, such styles are wrongly
inherited from the cache.

So this patch makes not to use cache for inehrited styles at the shadow boundary.

Tests: editing/shadow/breaking-editing-boundary-with-table.html
       editing/shadow/contenteditable-propagation-at-shadow-boundary.html

* css/StyleResolver.cpp:
(WebCore::StyleResolver::collectMatchingRulesForList):

LayoutTests:

Test cases to confirm that (1) selection does not break editing boundary,
and (2) content-editable is not propagated from shadow host to shadow DOM.

Actually a patch for Bug 88514 contains a test for (2), but it is not enough.

* editing/shadow/breaking-editing-boundary-with-table.html: Added.
* editing/shadow/contenteditable-propagation-at-shadow-boundary.html: Added.


Identifier: 107480@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120887 268f45cc-cd09-0410-ab3c-d52691b4dbfc